Make FastCGI the default server mode
It doesn't make a whole lot of sense for 'hhvm -m server' to throw an error
("HHVM no longer supports the built-in webserver as of 3.0.0.", etc.), and for
'hhvm -m server -vServer.Type=fastcgi' to be the correct invocation. HHVM
should assume you want the server implementation that exists, not the one that
doesn't. I've seen folks on #hhvm get stymied by this on at least two
occasions.

Also remove the WarmupDocument option; it doesn't do anything.