Finite chunk size for top-level output buffer
For HTTP requests, allow the transport to set the chunk size of the
top-level output buffer. This makes the buffer be peroidically flushed,
instead of growing indefinitely. With this change, Proxygen can stream
output indefinitely with O(1) memory, e.g. readfile('/dev/zero').
FastCGI needs a bit more work, which will be in a subsequent commit.