Use a separate flag for strict unflushable mode
PHP PHP interprets the flushable flag very loosely -- it only affects
ob_flush(), it doesn't affect ob_end_flush() or the automatic flush
which occurs when the request ends. I filed a bug for this in PHP but it
was refused. So instead, emulate PHP's behaviour, and introduce a new
flag OutputDisabled, which strictly prevents flushing and any other kind
of output from a buffer. It is similar to PHP's PHP_OUTPUT_DISABLED.

Instead of passing flags directly from userspace to obStart(), map them
to an enum class. This allows us to easily hide flags from userspace,
and also allows us to remove ext_std_output.h from execution-context.h.